program/requirements.txt
absl-py==2.2.2
albumentations==1.4.3
astunparse==1.6.3
cachetools==5.5.2
cellpose==3.1.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
csbdeep==0.8.1
cycler==0.12.1
efficientnet==1.0.0
efficientnet-pytorch==0.7.1
fastremap==1.15.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.30.2
idna==3.10
image-classifiers==1.0.0
imagecodecs==2025.3.30
imageio==2.37.0
Jinja2==3.1.6
joblib==1.4.2
keras==3.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
munch==4.0.0
namex==0.0.8
natsort==8.4.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pretrainedmodels==0.7.4
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
roifile==2025.2.20
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.3.2
scipy==1.15.2
segmentation-models==1.0.1
segmentation_models_pytorch==0.4.0
six==1.17.0
stardist==0.9.1
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-addons==0.16.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.15
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
